Add some (fairly loose) upper bounds on all dependencies
These packages are pretty foundational, and our dependency on them is
very minimal, so it’s unlikely that most updates will break them, but
the recent base16-bytestring-1.0.0.0 breakage highlights the dangers of
not even constraining the supermajor version. This adds a little more
safety without being terribly constraining.

Also, check in the .cabal file to play nicer with cabal-install.